Moved Galois hopper into internal namespace.
